refactor: move taggedHash to separate file
The tagged hash is defined in BIP0340 and also used in schnorr, so it
deserves a separate file.

Issue: BG-37835
