+++ src/taggedHash.js
@@ -1,0 +1,22 @@
+'use strict';
+Object.defineProperty(exports, '__esModule', { value: true });
+exports.taggedHash = void 0;
+const bcrypto = require('./crypto');
+const TAGS = [
+  'TapLeaf',
+  'TapBranch',
+  'TapTweak',
+  'KeyAgg list',
+  'KeyAgg coefficient',
+];
+/** An object mapping tags to their tagged hash prefix of [SHA256(tag) | SHA256(tag)] */
+const TAGGED_HASH_PREFIXES = Object.fromEntries(
+  TAGS.map(tag => {
+    const tagHash = bcrypto.sha256(Buffer.from(tag));
+    return [tag, Buffer.concat([tagHash, tagHash])];
+  }),
+);
+function taggedHash(prefix, data) {
+  return bcrypto.sha256(Buffer.concat([TAGGED_HASH_PREFIXES[prefix], data]));
+}
+exports.taggedHash = taggedHash;
