+++ ts_src/taggedHash.ts
@@ -1,0 +1,21 @@
+import * as bcrypto from './crypto';
+
+const TAGS = [
+  'TapLeaf',
+  'TapBranch',
+  'TapTweak',
+  'KeyAgg list',
+  'KeyAgg coefficient',
+] as const;
+type TaggedHashPrefix = typeof TAGS[number];
+/** An object mapping tags to their tagged hash prefix of [SHA256(tag) | SHA256(tag)] */
+const TAGGED_HASH_PREFIXES = Object.fromEntries(
+  TAGS.map(tag => {
+    const tagHash = bcrypto.sha256(Buffer.from(tag));
+    return [tag, Buffer.concat([tagHash, tagHash])];
+  }),
+) as { [k in TaggedHashPrefix]: Buffer };
+
+export function taggedHash(prefix: TaggedHashPrefix, data: Buffer): Buffer {
+  return bcrypto.sha256(Buffer.concat([TAGGED_HASH_PREFIXES[prefix], data]));
+}
